Added Disjunctive Faceting to the search method
The `search` method now accepts an additional option called
`disjunctiveFacets`. By denoting a facet as disjunctive, that means
when it returns facet counts, those counts will NOT account
for a value from this particular field being filtered upon.

The use case for this is something like tabs.

For instance, imagine a type "Cat" which has a field of "type".

In your UI, you may wish to have a "type" tab at the top of a page,
something like:

Happy (100) | Sad (1000) | Gray (500) | Black (200)

If you implement them as a traditional `facet`, selecting the
option "Black" would apply a filter on "type" of "Black". Your facet
counts on the next response would then be equivalent to:

Happy (0) | Sad (0) | Gray (0) | Black (200)

To avoid this, we simply do not account for the field of the
disjunctive facet if it is being filtered upon.