Fixed disjunctive facets.
In faceting, multi-value facet filters often send filters in the
following format:

```
{
  all: [
    {
      any: [
        { brand: 'a' },
        { brand: 'c' },
        { brand: 'c' }
      ]
    },
    {
      any: [
        { size: 'small' },
        { brand: 'medium' }
      ]
    }
  ]
}
```

When determining the fields that are currently being filterd, our
faceting logic was only looking at keys at the top level, so faceting
logic was not being applied when it should have. We are now recursing
down the filter tree to find values nested in this way.